src/labs/jwt/jwt_authentication_bypass_via_algorithm_confusion_with_no_exposed_key/pyproject.toml
[tool.poetry]
name = "standalone"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
asn1tools = "^0.155.3"
bitstruct = "^8.11.1"
cffi = "^1.16.0"
cryptography = "^3.3.1"
diskcache = "^5.2.1"
gmpy2 = "^2.2.1"
prompt-toolkit = "^3.0.14"
pycparser = "^2.20"
pycryptodome = "^3.20.0"
pyparsing = "^2.4.7"
six = "^1.15.0"
wcwidth = "^0.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"